+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+##  (2020-03-19)
+
+
+#### Features
+
+* **Volume:** mounted path and attach volumes by start ([82c50883](82c50883))
+
+#### Bug Fixes
+
+* **Linting:** blacked code ([9200bdf7](9200bdf7))
+* **request:**  add production boolean to config ([a083298a](a083298a))
 
 ##  (2020-03-13)